Fix a bug where stack objects could be overwritten
The problem was that LLVM assumed the spill size for scalar registers is 16 bits.  This in turn was because I set v16i1 as a valid type for ScalarRegister. Because it still used a 32 bit alignment, this generally didn't cause a problem, but if there were smaller objects on the stack, these could overwrite another stack slot.