issue #9: fix issue with CFI encoding
The PC ranges for each frame description entry was using the type
'DW_EH_PE_pcrel' by default if a target didn't override it. This
would end up trying to compute the address relative to the data
segment that contained .eh_frame, which would result in incorrect
addresses. Changed this to be absolute for Nyuzi.