issue #9: got stack crawls basically working
- In NyuziMCTargetDesc.cpp: when initializing NyuziMCRegisterInfo, pass
the proper return address register to the constructor. This was erroneously
passing the frame pointer before. This gets encoded into the CIE entry
in the .eh_frame section and used by the debugger to find the next function
- In NyuziFrameLowering.cpp, emit DW_CFA_def_cfa instead of DW_CFA_def_cfa_offset.
The former also specifies the canonical frame address register (CFA), which
is the stack pointer. Otherwise the debugger chokes.
- The register table in Plugins/ABINyuzi (in lldb) had some incorrect indices.
I don't think this is actually used, but fixed it anyway.
- Use the DWARF register number when setting up return address pointer for
default unwind plan.  Also not used in most cases.
- In LLDB, ABINyuzi, CallFrameAddressIsValid, it was using hex instead of
decimal to check for stack alignment, which would incorrectly flag some
CFAs as invalid.
- In LLDB, DWARFCallFrameInfo.cpp, FDEToUnwindPlan, the initial DW_CFA_advance_loc
instruction was creating an invalid row in the UnwindPlan. Added a hack to not
push a new row for the first instruction.