Fix issue #5 MCLinker crashes when using linker script
There was a version of FlexLexer.h checked into the mclinker tree.
Some parts of the build apparently were using the system include
and others were using the system version. The system version uses
size_t instead of int in some spots. size_t is now 64 bits on mac,
so the structure size and layout differs.

This object is stack allocated in ScriptReader::readScript, using
the smaller mclinker version of yyFlexLexer. The constructor
(which is compiled separately and uses the larger system version)
then overruns the stack.

The fix is just to remove FlexLexer from the mclinker tree.