Fix issue #42
The issue happened because the machine basic block iterator was at the end of the
block in NyuziInstrInfo::adjustStack, and the latter attempted to dereference it
to find the debug info.

This was actually wrong for another reason: NyuziFrameLowering::emitPrologue
is not supposed to emit debug info in the prologue, since the first debug location
is used to determine the end of the prologue.

Made debugInfo be passed into adjustStack as a parameter, uninitialized in emitPrologue.