issue #47: assertion trying to scavenge a register
Because each instruction's immediate field is only 13 bits, if a stack frame is
larger than 4k, the prolog inserter needs a temporary register to adjust the
stack pointer. As it does this after the register allocation, this requires the
register scavenger to run.

NyuziFrameLowering::determineCalleeSaves was treating the field as unsigned and
only allocating a save slot on the stack for the register scavenger if the
stack was more than 8k. So, if the stack frame was between 4k and 8k, and it
needed save a physical register value, it would assert.

This is difficult to write a simple unit test for because it's not sufficient
to only allocate a 4k stack frame, it also needs to save a register.