Fix assertions that would occur with LowerVECTOR_SHUFFLE
issue #57 and #58
When generating a new node, this function had hardcoded assumptions that the
result would always be a vector of integers. This could break instruction
matching in some cases.

This is a bit tricky to test this cleanly, because the failures only trigger
assertions as side effects of other types of transformations, so I didn't
create any regression tests for this.