Remove 'vector = scalar op immediate' instruction forms
This simplifies the instruction set and hardware implementation a bit.
As this is a fairly rare option, the performance increase isn't worth the
extra complexity.

But, more importantly, this frees up a bit from the type field, allowing
widening the immediate value field from 13 to 14 bits.  This in turn will
(in a subsequent commit) allow adding an instruction to load the upper
19 bits of a register, which allows loading a 32 bit value in two
instructions. My goal is to eliminate PC relative constant pool loads,
as they don't work when the frame is too large.