Switch to loadhi/or to get constant pool, block, and global addresses
This removes the previous hack of using PC relative loads and modifying EmitConstant
pool to emit the constant pool near the function that used it. The advantage of
the new approach is that it doesn't have size limitations and is a step towards
removing the PC from the register file.

This is slightly slower, because it takes 3 instructions to load a constant or
global instead of 1. This could be reduced to two instructions by using a movhi/load
combination, which I'll look into later.