Override isMachineVerifierClean and make it return false
In the last upstream sync, change 304320 was merged which
added this to all backends. The check-in comment was:

"This adds a callback to the LLVMTargetMachine that lets target indicate
that they do not pass the machine verifier checks in all cases yet.
This is intended to be a temporary measure while the targets are fixed
allowing us to enable the machine verifier by default with
EXPENSIVE_CHECKS enabled!"

I added this flag to Nyuzi to be consistent (although I saw no
problems with it off). As the other backends get fixed (whatever
that involves), I will look to modify Nyuzi accordingly.