Work towards issue #69: Support PIC relocation model
This change breaks assembly source level compatibility, as 's28' has
been renamed to 'gp' (global pointer).

Added basic global offset table support. This is not wired up
in the compiler, only the assembler. Can load from the global offset
table like this:

   load_32 s0, got(symname)(gp)

This will create a relocation. The linker supports emitting an executable
with the proper GOT entries. Currently, this form has a limit of 4096
entries, which is probably going to be overly restrictive. Will probably
add an extended GOT index form.