A few PIC fixes
- Set HasRelocationAddend to false when initializing MCELFObjectTargetWriter. This was
causing problems with GOT, because the linker was applying the addend to the GOT
index instead of the entry within the GOT. MIPs has hardcoded support for doing the
latter scattered all over the linker, but, to keep the changes contained and
simpler, I just disable addends alltogether.
- Make needsRelocateWithSymbol always return true. This is related to above. When this
was false, the backend would emit relocations that pointed to the segment base with
an addend. Then the code above would lose the addend. This causes every relocation to
point to the exact location it is targeting.