Fix jump table code generation with PIC enabled
When compiling in PIC code, each jump table entry needs a relocation
with an addend that represents the difference between the target and
the base of the jump table.

The problem was that LLVM was putting the jump table in a different segment
than the code (.rodata vs. .text), which made the addend meaningless
(since we can't know at that point what the section layout is). The default
implementation of TargetLoweringObjectFileELF returns false for
shouldPutJumpTableInFunctionSection by default, which doesn't make much
sense, since that can never work, and LLVM already went to the trouble
to determine if it needed to be in the same section.

Also, in NyuziELFObjectWriter, changed HasRelocationAddend to true. This
needs to be true to allow the added for jump tables to be emitted.
Modified needsRelocateWithSymbol so it only requires this for GOT
entries (GOT entries cannot use an addend, since it ends up being
applied to the GOT index instead of its target)

With this change, the test programs I ran (Doom, quakeview) compiled
with PIC run correctly. They're not fully position independent yet,
since I haven't implemented the PLT, but the implementation of
GOT seems to be correct.