Change dyn_cast<> to cast<> in a few places
In these instances, code was dereferencing the pointer without checking if it
was null. This was because the code couldn't work correctly if the type was
wrong (it would be a bug that this code was executing at all). the cast<>
will assert if the type is wrong, which would be the right thing to do.